fix(agent-mode): address codex review on background-task disabling
- Disallow Monitor alongside TaskOutput/Workflow: Monitor is inherently
  asynchronous with no foreground mode, so a watch could outlive the finite
  query and its result be lost with TaskOutput disabled (codex P2).
- Move the notification-only replay case into sdkMessageTranslator.test.ts
  under the translateSdkMessage suite so its coverage has one owner (codex P2).
- Resolve the replay fixture from __dirname instead of process.cwd() so the
  test is independent of Jest's working directory (codex P3).
